remote-agent/requirements.txt
python-socketio[client]==5.9.0
psutil==5.9.6
websocket-client==1.6.4
